Many new things.
Added a main menu scene (ALWAYS RUN FROM THIS SCENE UNTIL FURTHER NOTICE). Player, Main Camera, and UserInterfaceCanvas are now instantiated dynamically by the GameMaster script after selecting an option on the main menu. For these object scripts, moved code from Start to Awake since Start is not called immediately when Instantiate is used to create an object. References to Player, Main Camera, and UserInterfaceCanvas are now static to prevent multiple instances. When one of these objects is created in a scene, it checks to see if GameMaster already has an instance of itself. If it does, it immediately deletes itself to prevent conflicts.